feat(agentboard): replace composables with Pinia card store
Single source of truth for card state via useCardStore. WS events
update the store directly via bindWebSocket(). Deleted useCards and
useBoard composables.
