+++ plugins/tt-agentboard/app/plugins/init-store.client.ts
@@ -1,0 +1,6 @@
+export default defineNuxtPlugin(() => {
+  const cardStore = useCardStore();
+  const ws = useWebSocket();
+  cardStore.bindWebSocket(ws);
+  cardStore.fetchCards();
+});
