+++ plugins/tt-agentboard/package.json
@@ -14,10 +14,12 @@
     "test:watch": "vitest watch"
   },
   "dependencies": {
+    "@pinia/nuxt": "^0.11.3",
     "@vueuse/core": "^14.2.1",
     "chokidar": "^5.0.0",
     "consola": "^3.4.2",
     "nuxt": "^4.4.2",
+    "pinia": "^3.0.4",
     "vue": "^3.5.30",
     "vue-router": "^5.0.4",
     "vuedraggable": "^4.1.0",
